Phasen auswertung funktioniert

+++ TDD-Trainer/src/tracking/PhaseStartEvent.java
@@ -5,7 +5,7 @@ import java.time.temporal.ChronoUnit;
 
 public class PhaseStartEvent extends Event{
 	private int phase;
-	private long duration;
+	private int duration;
 	public PhaseStartEvent(int phase){
 		super();
 		this.phase = phase;
@@ -13,10 +13,10 @@ public class PhaseStartEvent extends Event{
 	public int getPhase(){
 		return phase;
 	}
-	public long getDuration(){
+	public int getDuration(){
 		return duration;
 	}
 	public void calculateDuration() {
-		duration =  super.time.until(LocalDateTime.now(), ChronoUnit.SECONDS);
+		duration =  (int)super.time.until(LocalDateTime.now(), ChronoUnit.SECONDS);
 	}
 }
